+++ boofuzz/cli_context.py
@@ -1,0 +1,10 @@
+import attr
+
+from .sessions import Session
+
+
+@attr.s
+class CliContext(object):
+    """Context for Click commands' Context.obj """
+
+    session = attr.ib(type=Session)